move Checkbox Delegate to utils
