+++ QgisModelBaker/utils/gui_utils.py
@@ -6,7 +6,9 @@ import warnings
 import xml.etree.ElementTree as CET
 from enum import Enum, IntEnum
 
+from PyQt5.QtWidgets import QApplication
 from qgis.PyQt.QtCore import (
+    QEvent,
     QModelIndex,
     QSortFilterProxyModel,
     QStringListModel,
@@ -14,7 +16,14 @@ from qgis.PyQt.QtCore import (
     pyqtSignal,
 )
 from qgis.PyQt.QtGui import QIcon, QStandardItem, QStandardItemModel
-from qgis.PyQt.QtWidgets import QCheckBox, QLineEdit, QListView
+from qgis.PyQt.QtWidgets import (
+    QCheckBox,
+    QLineEdit,
+    QListView,
+    QStyle,
+    QStyledItemDelegate,
+    QStyleOptionButton,
+)
 from qgis.PyQt.uic import loadUiType
 
 from QgisModelBaker.libs.modelbaker.iliwrapper.ilicache import IliCache
@@ -174,6 +183,7 @@ class PageIds:
     ExportDataConfiguration = 10
     ExportDataExecution = 11
     ProjectCreation = 12
+    TIDConfiguration = 13
 
 
 class ToppingWizardPageIds:
@@ -1107,3 +1117,23 @@ class BasketSourceModel(QStandardItemModel):
         for basket in self.schema_baskets[schema_identificator]:
             model_topics.add(basket.get("topic", ""))
         return list(model_topics)
+
+
+class CheckDelegate(QStyledItemDelegate):
+    def __init__(self, parent, role):
+        super().__init__(parent)
+        self.role = role
+
+    def editorEvent(self, event, model, option, index):
+        if event.type() == QEvent.MouseButtonRelease:
+            value = index.data(int(self.role)) or False
+            model.setData(index, not value, int(self.role))
+            return True
+        return super().editorEvent(event, model, option, index)
+
+    def paint(self, painter, option, index):
+        opt = QStyleOptionButton()
+        opt.rect = option.rect
+        value = index.data(int(self.role)) or False
+        opt.state |= QStyle.State_On if value else QStyle.State_Off
+        QApplication.style().drawControl(QStyle.CE_CheckBox, opt, painter)
